# public/swagger.json
{
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"SaveUpdateResAdmin": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"Name": {
						"type": "string"
					},
					"Email": {
						"type": "string"
					},
					"Password": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"Name",
					"Email",
					"Password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveReqAdmin": {
				"properties": {
					"Name": {
						"type": "string"
					},
					"Email": {
						"type": "string"
					},
					"Password": {
						"type": "string"
					}
				},
				"required": [
					"Name",
					"Email",
					"Password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateReqAdmin": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"Name": {
						"type": "string"
					},
					"Email": {
						"type": "string"
					},
					"Password": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"Name",
					"Email",
					"Password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LoginAdmin": {
				"properties": {
					"Email": {
						"type": "string"
					},
					"Password": {
						"type": "string"
					}
				},
				"required": [
					"Email",
					"Password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveUpdateResUser": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"PictureUrl": {
						"type": "string"
					},
					"Name": {
						"type": "string"
					},
					"UserName": {
						"type": "string"
					},
					"Password": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"PictureUrl",
					"Name",
					"UserName",
					"Password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveReqUSER": {
				"properties": {
					"PictureUrl": {
						"type": "string"
					},
					"Name": {
						"type": "string"
					},
					"UserName": {
						"type": "string"
					},
					"Password": {
						"type": "string"
					}
				},
				"required": [
					"PictureUrl",
					"Name",
					"UserName",
					"Password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateReqUSER": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"PictureUrl": {
						"type": "string"
					},
					"Name": {
						"type": "string"
					},
					"UserName": {
						"type": "string"
					},
					"Password": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"PictureUrl",
					"Name",
					"UserName",
					"Password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"_LeanDocument__LeanDocument_T__": {
				"properties": {},
				"type": "object"
			},
			"Pick__LeanDocument_T_.Exclude_keyof_LeanDocument_T_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested__": {
				"properties": {
					"_id": {
						"$ref": "#/components/schemas/_LeanDocument__LeanDocument_T__",
						"description": "This documents _id."
					},
					"__v": {
						"description": "This documents __v."
					},
					"id": {
						"description": "The string version of this documents _id."
					}
				},
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit__LeanDocument_this_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested_": {
				"$ref": "#/components/schemas/Pick__LeanDocument_T_.Exclude_keyof_LeanDocument_T_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"LeanDocument_this_": {
				"$ref": "#/components/schemas/Omit__LeanDocument_this_.Exclude_keyofDocument._id-or-id-or-__v_-or-%24isSingleNested_"
			},
			"IUSER": {
				"$ref": "#/components/schemas/LeanDocument_this_"
			},
			"DeleteUser": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LogInReqUSER": {
				"properties": {
					"UserName": {
						"type": "string"
					},
					"Password": {
						"type": "string"
					}
				},
				"required": [
					"UserName",
					"Password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MemberInterface": {
				"properties": {
					"UserId": {
						"type": "string"
					}
				},
				"required": [
					"UserId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Messageinterface": {
				"properties": {
					"User": {
						"type": "string"
					},
					"Message": {
						"type": "string"
					}
				},
				"required": [
					"User",
					"Message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveUpdateResgroup": {
				"properties": {
					"id": {
						"type": "string"
					},
					"Name": {
						"type": "string"
					},
					"Member": {
						"items": {
							"$ref": "#/components/schemas/MemberInterface"
						},
						"type": "array"
					},
					"Message": {
						"items": {
							"$ref": "#/components/schemas/Messageinterface"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"Name",
					"Member",
					"Message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SaveGruopReq": {
				"properties": {
					"Name": {
						"type": "string"
					}
				},
				"required": [
					"Name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IGROUP": {
				"$ref": "#/components/schemas/LeanDocument_this_"
			},
			"DeletGruopReq": {
				"properties": {
					"_id": {
						"type": "string"
					}
				},
				"required": [
					"_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AddUserInGruop": {
				"properties": {
					"id": {
						"type": "string"
					},
					"gruopId": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"gruopId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReturnGroupMessage": {
				"properties": {
					"_id": {
						"type": "string"
					}
				},
				"required": [
					"_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MessageInGroup": {
				"properties": {
					"groupId": {
						"type": "string"
					},
					"Messages": {
						"items": {
							"$ref": "#/components/schemas/Messageinterface"
						},
						"type": "array"
					}
				},
				"required": [
					"groupId",
					"Messages"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CheckingMessage": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "token",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "company",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/admin/saveadmin": {
			"post": {
				"operationId": "SaveAdmin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResAdmin"
								}
							}
						}
					}
				},
				"tags": [
					"admin"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SaveReqAdmin"
							}
						}
					}
				}
			}
		},
		"/admin/updateadmin": {
			"put": {
				"operationId": "UpdateAdmin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResAdmin"
								}
							}
						}
					}
				},
				"tags": [
					"admin"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateReqAdmin"
							}
						}
					}
				}
			}
		},
		"/admin/loginadmin": {
			"post": {
				"operationId": "LoginAdmin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResAdmin"
								}
							}
						}
					}
				},
				"tags": [
					"admin"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginAdmin"
							}
						}
					}
				}
			}
		},
		"/user/saveuser": {
			"post": {
				"operationId": "SaveUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResUser"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SaveReqUSER"
							}
						}
					}
				}
			}
		},
		"/user/updateUser": {
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResUser"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateReqUSER"
							}
						}
					}
				}
			}
		},
		"/user/deleteUser": {
			"delete": {
				"operationId": "Deletadmin",
				"responses": {
					"200": {
						"description": "User is deleted",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IUSER"
										},
										{
											"properties": {
												"_id": {
													"type": "string"
												}
											},
											"required": [
												"_id"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteUser"
							}
						}
					}
				}
			}
		},
		"/user/loginUser": {
			"post": {
				"operationId": "LoginUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResUser"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LogInReqUSER"
							}
						}
					}
				}
			}
		},
		"/group/savegrop": {
			"post": {
				"operationId": "SaveGrop",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResgroup"
								}
							}
						}
					}
				},
				"tags": [
					"Group"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SaveGruopReq"
							}
						}
					}
				}
			}
		},
		"/group/deletegrop": {
			"delete": {
				"operationId": "DeleteGrop",
				"responses": {
					"200": {
						"description": "User is deleted",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/IGROUP"
										},
										{
											"properties": {
												"_id": {
													"type": "string"
												}
											},
											"required": [
												"_id"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Group"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeletGruopReq"
							}
						}
					}
				}
			}
		},
		"/group/addmembertogroup": {
			"post": {
				"operationId": "AddMembertoGroup",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResgroup"
								}
							}
						}
					}
				},
				"tags": [
					"Group"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AddUserInGruop"
							}
						}
					}
				}
			}
		},
		"/group/returnmessages": {
			"post": {
				"operationId": "ReturnGropMessage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResgroup"
								}
							}
						}
					}
				},
				"tags": [
					"Group"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ReturnGroupMessage"
							}
						}
					}
				}
			}
		},
		"/group/savemessages": {
			"post": {
				"operationId": "SaveMessage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResgroup"
								}
							}
						}
					}
				},
				"tags": [
					"Group"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MessageInGroup"
							}
						}
					}
				}
			}
		},
		"/group/checkingMessage": {
			"post": {
				"operationId": "CheckingMesage",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SaveUpdateResgroup"
								}
							}
						}
					}
				},
				"tags": [
					"Group"
				],
				"security": [
					{
						"api_key": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CheckingMessage"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}